fix: restore main CI checks
