+++ src/system-agent/operations-parse.ts
@@ -1,7 +1,119 @@
-// Closed, anchored grammar for OpenClaw system-agent operations.
+// OpenClaw operation grammar, approval descriptions, and public types.
+import type { ConfigSetOptions } from "../cli/config-set-input.js";
+import type { DoctorOptions } from "../commands/doctor.types.js";
+import { isSensitiveConfigPath } from "../config/sensitive-paths.js";
 import { normalizeAgentId } from "../routing/session-key.js";
-import type { SystemAgentOperation } from "./operations.js";
+import type { RuntimeEnv } from "../runtime.js";
+import type { TuiResult } from "../tui/tui-types.js";
+import { resolveUserPath, shortenHomePath } from "../utils.js";
+import { isReservedSystemAgentId } from "./agent-id.js";
+import type { SystemAgentOverview } from "./overview.js";
 import { validateSystemAgentPluginInstallSpec } from "./plugin-install.js";
+
+type SystemAgentOverviewLoader = () => Promise<SystemAgentOverview>;
+type SystemAgentOverviewFormatter = (overview: SystemAgentOverview) => string;
+
+/** Parsed OpenClaw operation before approval/execution. */
+export type SystemAgentOperation =
+  | { kind: "none"; message: string }
+  | { kind: "overview" }
+  | { kind: "doctor" }
+  | { kind: "doctor-fix" }
+  | { kind: "status" }
+  | { kind: "health" }
+  | { kind: "config-validate" }
+  | { kind: "config-get"; path: string }
+  | { kind: "config-schema"; path?: string }
+  | { kind: "config-set"; path: string; value: string }
+  | {
+      kind: "config-set-ref";
+      path: string;
+      source: "env" | "file" | "exec";
+      id: string;
+      provider?: string;
+    }
+  | { kind: "setup"; workspace?: string; model?: string }
+  | { kind: "model-setup"; workspace?: string }
+  | { kind: "channel-list" }
+  | { kind: "channel-info"; channel: string }
+  | { kind: "channel-setup"; channel: string }
+  | {
+      kind: "open-setup";
+      target: "guided" | "classic" | "channels";
+      channel?: string;
+    }
+  | { kind: "gateway-status" }
+  | { kind: "gateway-start" }
+  | { kind: "gateway-stop" }
+  | { kind: "gateway-restart" }
+  | { kind: "agents" }
+  | { kind: "models" }
+  | { kind: "plugin-list" }
+  | { kind: "plugin-search"; query: string }
+  | { kind: "plugin-install"; spec: string }
+  | { kind: "plugin-uninstall"; pluginId: string }
+  | { kind: "audit" }
+  | { kind: "create-agent"; agentId: string; workspace?: string; model?: string }
+  | { kind: "open-tui"; agentId?: string; workspace?: string }
+  | { kind: "set-default-model"; model: string };
+
+/** Result returned by the operation executor. */
+export type SystemAgentOperationResult = {
+  applied: boolean;
+  exitsInteractive?: boolean;
+  message?: string;
+  nextInput?: string;
+  /** Agent TUI exited via /openclaw: re-enter the shell even without a request. */
+  returnToShell?: boolean;
+  followUp?: Extract<SystemAgentOperation, { kind: "model-setup" }>;
+};
+
+/** Injectable command dependencies used by tests and alternate runners. */
+export type SystemAgentCommandDeps = {
+  readConfigFileSnapshot?: typeof import("../config/config.js").readConfigFileSnapshot;
+  ensureAuthProfileStore?: typeof import("../agents/auth-profiles/store.js").ensureAuthProfileStore;
+  resolveCliAuthBindingFingerprint?: typeof import("../agents/cli-auth-epoch.js").resolveCliAuthBindingFingerprint;
+  resolveApiKeyForProvider?: typeof import("../agents/model-auth.js").resolveApiKeyForProvider;
+  formatOverview?: SystemAgentOverviewFormatter;
+  loadOverview?: SystemAgentOverviewLoader;
+  runAgentsAdd?: (
+    opts: {
+      name?: string;
+      workspace?: string;
+      model?: string;
+      nonInteractive?: boolean;
+      json?: boolean;
+    },
+    runtime: RuntimeEnv,
+    params?: { hasFlags?: boolean },
+  ) => Promise<void>;
+  runConfigSet?: (opts: {
+    path?: string;
+    value?: string;
+    cliOptions: ConfigSetOptions;
+  }) => Promise<void>;
+  runDoctor?: (runtime: RuntimeEnv, options: DoctorOptions) => Promise<void>;
+  runGatewayRestart?: () => Promise<void | boolean>;
+  runGatewayStart?: () => Promise<void>;
+  runGatewayStop?: () => Promise<void>;
+  runPluginInstall?: (spec: string, runtime: RuntimeEnv) => Promise<void>;
+  runPluginUninstall?: (pluginId: string, runtime: RuntimeEnv) => Promise<void>;
+  runPluginsList?: (runtime: RuntimeEnv) => Promise<void>;
+  runPluginsSearch?: (query: string, runtime: RuntimeEnv) => Promise<void>;
+  runTui?: (opts: {
+    local: boolean;
+    session?: string;
+    deliver?: boolean;
+    historyLimit?: number;
+  }) => Promise<TuiResult | void>;
+  /** Where setup side effects run; the gateway surface never manages its own daemon. */
+  setupSurface?: "cli" | "gateway";
+  applySetup?: typeof import("./setup-apply.js").applySystemAgentSetup;
+  verifyInferenceConfig?: typeof import("./setup-inference.js").verifySetupInferenceConfig;
+  listChannelSetupPlugins?: typeof import("../channels/plugins/setup-registry.js").listChannelSetupPlugins;
+  resolveChannelSetupEntries?: typeof import("../commands/channel-setup/discovery.js").resolveChannelSetupEntries;
+  isChannelConfigured?: typeof import("../config/channel-configured-shared.js").isStaticallyChannelConfigured;
+};
 
 // Grammar tokens. Workspace/path tokens accept quoted strings so paths with
 // spaces survive; model refs and ids stay single tokens.
@@ -282,3 +394,73 @@ function normalizePluginInstallSpec(spec: string, source: string | undefined): s
  * Guided setup operations are intentionally absent: starting a wizard is not
  * itself a write; the wizard owns approval and persistence for its answers.
  */
+export function isPersistentSystemAgentOperation(operation: SystemAgentOperation): boolean {
+  return (
+    operation.kind === "set-default-model" ||
+    operation.kind === "config-set" ||
+    operation.kind === "config-set-ref" ||
+    operation.kind === "setup" ||
+    operation.kind === "plugin-install" ||
+    (operation.kind === "create-agent" &&
+      !operation.model?.trim() &&
+      !isReservedSystemAgentId(operation.agentId)) ||
+    operation.kind === "gateway-start" ||
+    operation.kind === "gateway-stop" ||
+    operation.kind === "gateway-restart"
+  );
+}
+
+/** Format a user-facing description for an operation requiring approval. */
+export function describeSystemAgentPersistentOperation(operation: SystemAgentOperation): string {
+  switch (operation.kind) {
+    case "set-default-model":
+      return `set agents.defaults.model.primary to ${operation.model}`;
+    case "config-set":
+      return `set config ${operation.path} to ${formatConfigSetValueForPlan(operation.path, operation.value)}`;
+    case "config-set-ref":
+      return `set config ${operation.path} to ${operation.source} SecretRef ${operation.source === "env" ? operation.id : "<redacted>"}`;
+    case "setup":
+      return formatSetupPlanDescription(operation);
+    case "model-setup":
+      return "configure a model provider and default model";
+    case "doctor-fix":
+      return "exit OpenClaw and run openclaw doctor --fix";
+    case "plugin-install":
+      return `install plugin ${operation.spec}`;
+    case "plugin-uninstall":
+      return `uninstall plugin ${operation.pluginId}`;
+    case "create-agent":
+      return `create agent ${operation.agentId} with workspace ${formatCreateAgentWorkspace(operation.workspace)}`;
+    case "gateway-start":
+      return "start the Gateway";
+    case "gateway-stop":
+      return "stop the Gateway";
+    case "gateway-restart":
+      return "restart the Gateway";
+    default:
+      return "apply this action";
+  }
+}
+
+/** Format the standard approval plan text for a persistent operation. */
+export function formatSystemAgentPersistentPlan(operation: SystemAgentOperation): string {
+  return `Plan: ${describeSystemAgentPersistentOperation(operation)}. Say yes to apply.`;
+}
+
+function formatCreateAgentWorkspace(workspace: string | undefined): string {
+  return workspace ? shortenHomePath(resolveUserPath(workspace)) : shortenHomePath(process.cwd());
+}
+
+function formatConfigSetValueForPlan(configPath: string, value: string): string {
+  if (isSensitiveConfigPath(configPath)) {
+    return "<redacted>";
+  }
+  return value;
+}
+
+function formatSetupPlanDescription(
+  operation: Extract<SystemAgentOperation, { kind: "setup" }>,
+): string {
+  const workspace = shortenHomePath(resolveUserPath(operation.workspace ?? process.cwd()));
+  return `bootstrap OpenClaw setup for workspace ${workspace}`;
+}
